+++ src/tools/qencounterdelegate.cpp
@@ -12,7 +12,7 @@ QWidget *QEncounterDelegate::createEditor(QWidget *parent, const QStyleOptionVie
     Q_UNUSED(index)
     Q_UNUSED(option)
     QComboBox *editor = new QComboBox(parent);
-    for (int i = 0; i < (int)Data::troops.size(); i++)
+    for (size_t i = 0; i < Data::troops.size(); i++)
         editor->addItem(QString::fromStdString(Data::troops[i].name));
     return editor;
 }
@@ -27,7 +27,7 @@ void QEncounterDelegate::setModelData(QWidget *editor, QAbstractItemModel *model
 {
     int id = static_cast<QComboBox*>(editor)->currentIndex() + 1;
     model->setData(index, id, Qt::UserRole);
-    model->setData(index, QString::fromStdString(Data::troops[id-1].name));
+    model->setData(index, QString::fromStdString(Data::troops[static_cast<size_t>(id)-1].name));
 }
 
 void QEncounterDelegate::updateEditorGeometry(QWidget *editor, const QStyleOptionViewItem &option, const QModelIndex &index) const
